Add locale support to utility functions in format-numbers.js
